+++ pkg/apis/linstor/group.go
@@ -15,21 +15,9 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package controller
-
-import (
-	"sigs.k8s.io/controller-runtime/pkg/manager"
-)
-
-// AddToManagerFuncs is a list of functions to add all Controllers to the Manager
-var AddToManagerFuncs []func(manager.Manager) error
-
-// AddToManager adds all Controllers to the Manager
-func AddToManager(m manager.Manager) error {
-	for _, f := range AddToManagerFuncs {
-		if err := f(m); err != nil {
-			return err
-		}
-	}
-	return nil
-}
+// Package linstor contains linstor API versions.
+//
+// This file ensures Go source parsers acknowledge the linstor package
+// and any child packages. It can be removed if any other Go source files are
+// added to this package.
+package linstor