+++ pkg/apis/linstor/v1alpha1/doc.go
@@ -1,5 +1,5 @@
 /*
-Knopic Operator
+Linstor Operator
 Copyright 2019 LINBIT USA, LLC.
 
 Licensed under the Apache License, Version 2.0 (the "License");
@@ -15,9 +15,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-// Package knopic contains knopic API versions.
-//
-// This file ensures Go source parsers acknowledge the knopic package
-// and any child packages. It can be removed if any other Go source files are
-// added to this package.
-package knopic
+// Package v1alpha1 contains API Schema definitions for the linstor v1alpha1 API group
+// +k8s:deepcopy-gen=package,register
+// +groupName=linstor.linbit.com
+package v1alpha1