+++ CHANGELOG.md
@@ -19,14 +19,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - examples/operator-intra.yaml file to bundle all the rbac, crds, etc to run the
   operator
-- EtcdURL field to controllersetcontroller spec. default: etcd-knopic:2379
+- EtcdURL field to controllersetcontroller spec. default: etcd-linstor:2379
 - Host networking on the LINSTOR Satellite pods with DNSClusterFirstWithHostNet
   DNS policy
 - NodeSet service for the Satellite pods that also point to the Controller 
   service for LINSTOR discovery
 
 ### Removed
-- `ControllerEndpoint` and `DefaultController` from the KnopicNodeSet spec
+- `ControllerEndpoint` and `DefaultController` from the LinstorNodeSet spec
 
 ### Changed
 - Controller persistence is now handled by etcd. There must be a reachable and
@@ -34,7 +34,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Networking is now handled by a kubernetes service with the same name
   as the ControllerSet. The NodeSet must have the same name as the ControllerSet
   for networking to work properly.
-- Opt-in node label for nodes is now `linstor.linbit.com/knopic-node=true`
+- Opt-in node label for nodes is now `linstor.linbit.com/linstor-node=true`
 - Remove requirement for `kube-system` namespace
 - Naming convention for NodeSet and ControllerSet Pods
 - Updated ports for LINSTOR access on NodeSet and ControllerSet pods